Update BLM app to use threshold as float

+++ src/l2Mps_blm.h
@@ -87,16 +87,16 @@ public:
     const bool      getThresholdMaxEn   ( const blmThr_channel_t& ch) const { return findThrChannel(ch.blmCh)->getThresholdMaxEn(ch.thrTb); };
 
     // Threshold set methods
-    void            setThresholdMin     ( const blmThr_channel_t& ch, const uint32_t val) const { findThrChannel(ch.blmCh)->setThresholdMin(ch.thrTb, val); };
-    void            setThresholdMax     ( const blmThr_channel_t& ch, const uint32_t val) const { findThrChannel(ch.blmCh)->setThresholdMax(ch.thrTb, val); };
+    void            setThresholdMin     ( const blmThr_channel_t& ch, const float val) const { findThrChannel(ch.blmCh)->setThresholdMin(ch.thrTb, val); };
+    void            setThresholdMax     ( const blmThr_channel_t& ch, const float val) const { findThrChannel(ch.blmCh)->setThresholdMax(ch.thrTb, val); };
        
     // Threshold get methods
-    const uint32_t  getThresholdMin     ( const blmThr_channel_t& ch) const { return findThrChannel(ch.blmCh)->getThresholdMin(ch.thrTb); };
-    const uint32_t  getThresholdMax     ( const blmThr_channel_t& ch) const { return findThrChannel(ch.blmCh)->getThresholdMax(ch.thrTb); };
+    const float     getThresholdMin     ( const blmThr_channel_t& ch) const { return findThrChannel(ch.blmCh)->getThresholdMin(ch.thrTb); };
+    const float     getThresholdMax     ( const blmThr_channel_t& ch) const { return findThrChannel(ch.blmCh)->getThresholdMax(ch.thrTb); };
 
     // Scale factors set/get methods for thresholds 
     void            setScaleFactor      ( const blm_channel_t& ch, const float sf) const { findThrChannel(ch)->setScaleFactor(sf);          };
-    const float     getScaleFactor      ( const blm_channel_t& ch) const            { return findThrChannel(ch)->getScaleFactor();     };
+    const float     getScaleFactor      ( const blm_channel_t& ch) const                 { return findThrChannel(ch)->getScaleFactor();     };
 
     // Set polling thread with callback function
     const void      startPollThread     ( unsigned int poll, p_func_t callBack );
